# internal/route/web/assets/openapi.yaml
openapi: 3.1.0
info:
  version: 0.7.0
  title: Leagueify API
  summary: Open Source Sporting League Platform.
  description: Host your own sports league using Leagueify, the all-in-one league hosting platform.
  license:
    name: MIT
    identifier: MIT
    url: https://raw.githubusercontent.com/Leagueify/leagueify/main/LICENSE
...
servers:
  - url: http://localhost/api
    description: Dockerized Local Environment
  - url: http://localhost:8888/api
    description: Standalone Executable Local Environment
paths:
...
  /divisions:
    get:
      tags:
      - Divisions
      summary: List Divisions
      description: ''
      responses:
        200:
          $ref: "#/components/success/seasons"
        404:
          $ref: "#/components/errors/not-found"
    post:
      tags:
      - Divisions
      security:
        - bearerAuth: []
      summary: Create Divisions
      description: ''
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/division/creation"
      responses:
        201:
          $ref: "#/components/success/created"
        400:
          $ref: "#/components/errors/badRequest"
        401:
          $ref: "#/components/errors/unauthorized"
        403:
          $ref: "#/components/errors/forbidden"
...
components:
...
  division:
    creation:
      type: object
      properties:
        divisions:
          type: array
          description: Division Objects
          items:
            type: object
            required:
              - name
              - age
            properties:
              name:
                description: Division Name
                type: string
                example: Leagueify 18U
                minLength: 3
                maxLength: 32
              age:
                description: Division Age Object
                type: object
                required:
                  - min
                  - max
                properties:
                  min:
                    description: Minimum age for division
                    type: integer
                    example: 13
                  max:
                    description: Maximum age for division
                    type: integer
                    example: 18
...
  errors:
    badRequest:
      description: Error with request payload
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                description: HTTP status text
                type: string
                example: bad request
              detail:
                description: Details related to request payload error
                type: string
                example: invalid json payload
    unauthorized:
      description: Unauthorized response
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                description: HTTP status text
                type: string
                example: unauthorized
    forbidden:
      description: forbidden response
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                description: HTTP status text
                type: string
                example: forbidden
    not-found:
      description: not found reponse
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                description: HTTP status text
                type: string
                example: not found
  # successful responses
  success:
    created:
      description: Successfully created
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                description: HTTP status text
                type: string
                example: created
...
    seasons:
      description: List of League Seasons
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                description: HTTP status text
                type: string
                example: ok
              detail:
                description: List of league season ids and names
                type: array
                items:
                  description: Season objects
                  type: object
                  properties:
                    id:
                      description: Season ID
                      type: string
                      example: C329VF
                    name:
                      description: Season Name
                      type: string
                      example: Leagueify 2024
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT